23/416 refactoring add conditional render run npm run build
