+++ src/App.js
@@ -1,8 +1,8 @@
 import React,{Component} from "react";
 import 'tachyons';
-import CardList from "./CardList";
- import SearchBox from './SearchBox'
- import Scroll from "./scroll";
+import CardList from "./components/CardList";
+ import SearchBox from './components/SearchBox'
+ import Scroll from "./components/scroll";
 
 class App extends Component{
     constructor(){
@@ -28,16 +28,12 @@ class App extends Component{
         this.setState({searchfield: event.target.value})}
     
 render(){
-    const filteredRobots = this.state.robots.filter(robots=>{
-        return robots.name.toLowerCase().includes(this.state.searchfield.toLowerCase())
+    const {robots,searchfield} = this.state;
+    const filteredRobots = robots.filter(robots=>{
+        return robots.name.toLowerCase().includes(searchfield.toLowerCase())
     })
     console.log("render");
-    if(this.state.robots.length===0){
-        return <h1>Loading...</h1>
-    }
-    else{
-    return (
-        <div className='tc'>
+     return !robots.length ? <h1>Loading...</h1> :(<div className='tc'>
         <h1>RobotsFriends</h1>
         <SearchBox searchChange={this.onSearchChange}/>
         <Scroll>
@@ -45,7 +41,7 @@ render(){
         </Scroll>
         </div>
     );
-    }
+
 }
 }
